update sveltekit example to @sveltejs/kit@1.0.0-next.506 (release candidate)
