+++ examples/sveltekit-rc/.gitignore
@@ -5,3 +5,6 @@ node_modules
 /package
 .env
 .env.*
+!.env.example
+.vercel
+.output
