+++ examples/sveltekit-rc/src/app.d.ts
@@ -1,0 +1,4 @@
+// See https://kit.svelte.dev/docs/types#app
+// for information about these interfaces
+// and what to do when importing types
+declare namespace App {}
