+++ examples/sveltekit-rc/src/routes/+page.ts
@@ -1,0 +1,3 @@
+// since there's no dynamic data here, we can prerender
+// it so that it gets served as a static asset in production
+export const prerender = true;
