+++ examples/sveltekit-rc/src/routes/Counter.svelte
