+++ examples/sveltekit-rc/src/routes/about/+page.svelte
@@ -1,0 +1,3 @@
+<div><strong>/about</strong> route</div>
+<br>
+<a href="/" class="text-blue-600">Go Home</a>
