--- examples/sveltekit/.npmrc
@@ -1,1 +1,0 @@
-shamefully-hoist=true
