--- examples/sveltekit/src/app.html
@@ -1,11 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-	<head>
-		<meta charset="utf-8" />
-		<meta name="viewport" content="width=device-width, initial-scale=1" />
-		%sveltekit.head%
-	</head>
-	<body>
-		<div id="svelte">%sveltekit.body%</div>
-	</body>
-</html>
