--- examples/sveltekit/src/routes/about.svelte
@@ -1,6 +1,0 @@
-<script>
-  import 'uno.css'
-</script>
-<div><strong>/about</strong> route</div>
-<br>
-<a href="/" class="text-blue-600">Go Home</a>
