--- examples/sveltekit/src/routes/hi/[name].svelte
@@ -1,14 +1,0 @@
-<script lang="ts">
-	import 'uno.css'
-	import { page } from '$app/stores'
-
-	$: name = decodeURIComponent($page.params.name || '')
-</script>
-
-<div><strong>/hi</strong> route</div>
-
-<p>
-	Hi: { name }
-</p>
-<br>
-<a href="/" class="text-blue-600">Go Home</a>
