--- examples/sveltekit/tsconfig.json
@@ -1,3 +1,0 @@
-{
-  "extends": "./.svelte-kit/tsconfig.json"
-}
